Moved dummy executor into tests
